buildcache: continue swallowing fetch errors on copy
Since we have to try all the extensions for spec files, we need to continue
ignoring when some cannot be fetched.  Also, for "spack buildcache sync",
re-order the extensions we try so we start with most likely and try the least
likely last.

+++ lib/spack/spack/cmd/buildcache.py
@@ -17,6 +17,7 @@ import spack.cmd
 import spack.cmd.common.arguments as arguments
 import spack.config
 import spack.environment as ev
+import spack.fetch_strategy as fs
 import spack.hash_types as ht
 import spack.mirror
 import spack.relocate
@@ -635,10 +636,14 @@ def copy_buildcache_file(src_url, dest_url, local_path=None):
             temp_stage.create()
             temp_stage.fetch()
             web_util.push_to_url(local_path, dest_url, keep_original=True)
+        except fs.FetchError as e:
+            # Expected, since we have to try all the possible extensions
+            tty.debug("no such file: {0}".format(src_url))
+            tty.debug(e)
         finally:
             temp_stage.destroy()
     finally:
-        if tmpdir:
+        if tmpdir and os.path.exists(tmpdir):
             shutil.rmtree(tmpdir)
 
 
@@ -719,8 +724,9 @@ def sync_fn(args):
         buildcache_rel_paths.extend(
             [
                 os.path.join(build_cache_dir, bindist.tarball_path_name(s, ".spack")),
-                os.path.join(build_cache_dir, bindist.tarball_name(s, ".spec.yaml")),
+                os.path.join(build_cache_dir, bindist.tarball_name(s, ".spec.json.sig")),
                 os.path.join(build_cache_dir, bindist.tarball_name(s, ".spec.json")),
+                os.path.join(build_cache_dir, bindist.tarball_name(s, ".spec.yaml")),
             ]
         )
